build: replace lodash & clean up service-worker.js dependency
