+++ package.json
@@ -56,6 +56,7 @@
     "dayjs": "^1.11.13",
     "ecpair": "^3.0.0-rc.0",
     "ed25519-hd-key": "^1.3.0",
+    "es-toolkit": "^1.39.8",
     "ethereumjs-util": "^7.1.5",
     "ethers": "^6.13.4",
     "ethers-multicall-provider": "^6.5.0",
@@ -64,7 +65,6 @@
     "immer": "^10.1.1",
     "joi": "^17.13.3",
     "js-yaml": "^4.1.0",
-    "lodash": "^4.17.21",
     "long": "^5.2.3",
     "protobufjs": "^6.11.2",
     "qrcode.react": "^4.2.0",
@@ -94,7 +94,6 @@
     "@types/crypto-js": "^4.2.2",
     "@types/firefox-webext-browser": "^120.0.4",
     "@types/js-yaml": "^4.0.9",
-    "@types/lodash": "^4.17.13",
     "@types/long": "^4.0.2",
     "@types/node": "^22.7.4",
     "@types/qrcode.react": "^1.0.5",
@@ -117,6 +116,7 @@
     "typescript": "^5.5.4",
     "typescript-eslint": "^8.8.1",
     "vite": "^5.4.8",
+    "vite-plugin-es-toolkit": "^1.0.1",
     "vite-plugin-node-polyfills": "^0.22.0",
     "vite-plugin-static-copy": "^2.0.0",
     "vite-plugin-svgr": "^4.2.0",
