+++ src/hooks/evm/useAccountTxs.ts
@@ -1,4 +1,4 @@
-import { throttle } from 'lodash';
+import { throttle } from 'es-toolkit';
 
 import { MINTSCAN_FRONT_API_V10_URL } from '@/constants/common';
 import type { AccountTxsPayload } from '@/types/evm/txs';
