+++ src/hooks/sui/useTransactionBlocks.ts
@@ -1,4 +1,4 @@
-import { throttle } from 'lodash';
+import { throttle } from 'es-toolkit';
 import type { SuiTransactionBlockResponseQuery } from '@mysten/sui/client';
 
 import type { SuiRpcGetTransactionBlocksResponse } from '@/types/sui/api';
